backend completed frontend setup started

+++ backend/controllers/goalController.js
@@ -1,8 +1,9 @@
 const asyncHandler = require('express-async-handler')
 const Goal = require('../models/goalModel')
+const User = require('../models/userModel')
 
 const getGoals = asyncHandler(async (req,res) => {
-    const goals = await Goal.find()
+    const goals = await Goal.find({user : req.user.id})
     res.status(200).json(goals);
 })
 
@@ -12,7 +13,8 @@ const setGoal = asyncHandler(async (req,res) => {
         throw new Error('Please add a text field')
     }
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id,
     })
     res.status(200).json(goal);
 })
@@ -23,6 +25,19 @@ const updateGoal = asyncHandler(async (req,res) => {
         res.status(400)
         throw new Error('Goal not found')
     }
+
+    const user = await User.findById(req.user.id)
+
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    if(goal.user.toString()!== user.id){
+        res.status(401)
+        throw new Error('User not authorized')
+    }
+
     const updatedGoal=await Goal.findByIdAndUpdate(req.params.id, req.body, {
         new : true,
     })
@@ -30,7 +45,23 @@ const updateGoal = asyncHandler(async (req,res) => {
 })
 
 const deleteGoal = asyncHandler(async (req,res) => {
-    const { id } = req.params.id;
+    const goal = await Goal.findById(req.params.id)
+    if(!goal){
+        res.status(400)
+        throw new Error('Goal not found')
+    }
+
+    const user = await User.findById(req.user.id)
+
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    if(goal.user.toString()!== user.id){
+        res.status(401)
+        throw new Error('User not authorized')
+    }
     await Goal.deleteOne({ _id: req.params.id });
     res.status(200).json({id : req.params.id});
 })
